pyproject.toml
[project]
name = "42api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord>=2.3.2",
    "python-dotenv>=1.1.0",
    "requests-oauthlib>=2.0.0",
]
